\name{logopt-package}
\alias{logopt-package}
\alias{logopt}
\docType{package}
\title{ Universal log-optimal portfolios. }
\description{
  The package contains implementations in R for some universal portfolio
  strategies.
}
\details{
\tabular{ll}{
Package: \tab logopt\cr
Type: \tab Package\cr
Version: \tab 0.0-0\cr
Date: \tab 2010-10-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

The package contains series of related functions, based on published algorithms
}
\author{
  Marc Delvaux <mdelvaux@gmail.com>

Maintainer: Marc Delvaux <mdelvaux@gmail.com>
}
