\name{nyse.gyorfi.1962.2006}
\docType{data}
\alias{nyse.gyorfi.1962.2006}
\title{Some NYSE stocks as used by Gyorfi}
\description{
This data set is a xts time series of close to close ratios for some subset of NYSE stocks, extending in time the period used by Cover.
Some stocks have been dropped because they don't cover the full period, illustrating the danger of survivorship bias.
}
\usage{nyse.gyorfi.1962.2006}
\format{A xts time series, column names identify the stocks, plus Cash and Bond.}
\source{Available online at http://www.szit.bme.hu/~oti/portfolio/data.html}
\references{
Cover, Universal Portfolios
}
\keyword{datasets}