\name{nyse.cover.1962.1984}
\docType{data}
\alias{nyse.cover.1962.1984}
\title{Some NYSE stocks as used by Cover}
\description{
This data set is a xts time series of close to close ratios for some subset of NYSE stocks as originally used by Cover.
The same series was used as reference by some other articles.
}
\usage{nyse.cover.1962.1984}
\format{A xts time series, column names identify the stocks, plus Cash and Bond.}
\source{Available online at http://www.szit.bme.hu/~oti/portfolio/data.html }
\references{
Cover, Universal Portfolios
}
\keyword{datasets}